Add array access

+++ src/ast/expr.rs
@@ -92,7 +92,7 @@ pub enum Expr {
 pub enum GetAccessor {
 	DotName(Rc<str>),
 	DotEval(Box<Expr>),
-	SubscriptionNumber(usize),
+	SubscriptionNumber(f64),
 	SubscriptionEval(Box<Expr>),
 }
